+++ Netkan/Model/Metadata.cs
@@ -1,6 +1,10 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.ComponentModel;
+using System.Reflection;
 
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using YamlDotNet.RepresentationModel;
 
@@ -10,123 +14,126 @@ using CKAN.NetKAN.Transformers;
 
 namespace CKAN.NetKAN.Model
 {
+    [JsonObject(MemberSerialization.OptIn)]
     internal sealed class Metadata
     {
-        private const string KrefPropertyName          = "$kref";
-        private const string VrefPropertyName          = "$vref";
-        public  const string SpecVersionPropertyName   = "spec_version";
-        private const string VersionPropertyName       = "version";
-        private const string DownloadPropertyName      = "download";
-        public  const string UpdatedPropertyName       = "release_date";
-        private const string StagedPropertyName        = "x_netkan_staging";
-        private const string StagingReasonPropertyName = "x_netkan_staging_reason";
-
-        private readonly JObject _json;
-
-        public string         Identifier      => (string?)_json["identifier"] ?? "";
-        public RemoteRef?     Kref            { get; private set; }
-        public RemoteRef?     Vref            { get; private set; }
-        public ModuleVersion? SpecVersion     { get; private set; }
-        public ModuleVersion? Version         { get; private set; }
-        public Uri?           Download        { get; private set; }
-        public DateTime?      RemoteTimestamp { get; private set; }
-        public bool           Staged          { get; private set; }
-        public string?        StagingReason   { get; private set; }
-
-        public string[]       Hosts
-            => (_json.TryGetValue(DownloadPropertyName, out JToken? downloadToken)
-                    ? downloadToken.Type == JTokenType.String
-                      && (string?)downloadToken is string url
-                          ? Enumerable.Repeat(url, 1)
-                          : downloadToken.Children()
-                                         .Select(ch => (string?)ch)
-                                         .OfType<string>()
-                    : Enumerable.Empty<string>())
-                .Select(u => new Uri(u).Host)
-                .ToArray();
-
         public Metadata(JObject json)
         {
-            _json = json;
-
-            if (json.TryGetValue(KrefPropertyName, out JToken? krefToken))
+            AllJson = json;
+            try
             {
-                if (krefToken.Type == JTokenType.String)
-                {
-                    Kref = new RemoteRef((string?)krefToken ?? "");
-                }
-                else
-                {
-                    throw new Kraken(string.Format("{0} must be a string.", KrefPropertyName));
-                }
+                JsonSerializer.CreateDefault(UtcSettings)
+                              .Populate(json.CreateReader(), this);
             }
-
-            if (json.TryGetValue(VrefPropertyName, out JToken? vrefToken))
+            catch (TargetInvocationException exc)
             {
-                if (vrefToken.Type == JTokenType.String)
-                {
-                    Vref = new RemoteRef((string?)vrefToken ?? "");
-                }
-                else
-                {
-                    throw new Kraken(string.Format("{0} must be a string.", VrefPropertyName));
-                }
+                throw exc.GetBaseException();
             }
-
-            if (json.TryGetValue(SpecVersionPropertyName, out JToken? specVersionToken))
+            catch (JsonException jsonExc)
             {
-                if (specVersionToken.Type == JTokenType.Integer && (int)specVersionToken == 1)
-                {
-                    SpecVersion = new ModuleVersion("v1.0");
-                }
-                else if (specVersionToken.Type == JTokenType.String)
-                {
-                    SpecVersion = new ModuleVersion((string?)specVersionToken ?? "");
-                }
-                else
-                {
-                    throw new Kraken(string.Format(@"Could not parse {0}: ""{1}""",
-                        SpecVersionPropertyName,
-                        specVersionToken
-                    ));
-                }
-            }
-
-            if (json.TryGetValue(VersionPropertyName, out JToken? versionToken))
-            {
-                Version = new ModuleVersion((string?)versionToken ?? "");
-            }
-
-            if (json.TryGetValue(DownloadPropertyName, out JToken? downloadToken))
-            {
-                Download = (downloadToken.Type == JTokenType.String
-                                ? (string?)downloadToken
-                                : (string?)downloadToken.Children().First())
-                            is string s
-                                ? new Uri(s)
-                                : null;
-            }
-
-            if (json.TryGetValue(StagedPropertyName, out JToken? stagedToken))
-            {
-                Staged = (bool)stagedToken;
-            }
-
-            if (json.TryGetValue(StagingReasonPropertyName, out JToken? stagingReasonToken))
-            {
-                StagingReason = (string?)stagingReasonToken ?? "";
-            }
-
-            if (json.TryGetValue(UpdatedPropertyName, out JToken? updatedToken))
-            {
-                DateTime t = (DateTime)updatedToken;
-                RemoteTimestamp = t.ToUniversalTime();
+                throw new BadMetadataKraken(null, jsonExc.Message, jsonExc);
             }
         }
 
-        public Metadata(YamlMappingNode yaml) : this(yaml.ToJObject())
+        public Metadata(YamlMappingNode yaml)
+            : this(yaml.ToJObject())
         {
         }
+
+        [JsonProperty("identifier", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate,
+                                    NullValueHandling    = NullValueHandling.Ignore)]
+        [DefaultValue("")]
+        public readonly string Identifier = "";
+
+        [JsonProperty("$kref", NullValueHandling = NullValueHandling.Ignore)]
+        public readonly RemoteRef? Kref;
+
+        [JsonProperty("$vref", NullValueHandling = NullValueHandling.Ignore)]
+        public readonly RemoteRef? Vref;
+
+        [JsonProperty("spec_version", NullValueHandling = NullValueHandling.Ignore)]
+        public readonly ModuleVersion? SpecVersion;
+
+        [JsonProperty("version", NullValueHandling = NullValueHandling.Ignore)]
+        public readonly ModuleVersion? Version;
+
+        [JsonProperty("x_netkan_epoch", NullValueHandling = NullValueHandling.Ignore)]
+        public readonly uint? Epoch;
+
+        [JsonProperty("license")]
+        [JsonConverter(typeof(JsonSingleOrArrayConverter<License>))]
+        public readonly List<License>? Licenses;
+
+        public bool Redistributable => Licenses?.Any(lic => lic.Redistributable)
+                                               ?? false;
+
+        [JsonProperty("download", NullValueHandling = NullValueHandling.Ignore)]
+        [JsonConverter(typeof(JsonSingleOrArrayConverter<Uri>))]
+        public readonly List<Uri>? Download;
+
+        public Uri? FallbackDownload => Identifier != null
+                                     && Redistributable
+                                     && Version?.ToString().Replace(':', '-') is string verStr
+                                     && (JObject?)AllJson["download_hash"] is JObject hashes
+                                     && (string?)hashes["sha1"] is string sha1
+                                         ? new Uri($"https://archive.org/download/{Identifier}-{verStr}/{sha1[..8]}-{Identifier}-{verStr}.zip")
+                                         : null;
+
+        public string[] Hosts => Download?.Select(u => u.Host)
+                                          .ToArray()
+                                         ?? Array.Empty<string>();
+
+        [JsonProperty("release_date", NullValueHandling = NullValueHandling.Ignore)]
+        public readonly DateTime? ReleaseDate;
+
+        [JsonProperty("release_status", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate,
+                                        NullValueHandling    = NullValueHandling.Ignore)]
+        [DefaultValue(CKAN.ReleaseStatus.stable)]
+        public readonly ReleaseStatus? ReleaseStatus = CKAN.ReleaseStatus.stable;
+
+        public bool Prerelease => ReleaseStatus is CKAN.ReleaseStatus.development
+                                                or CKAN.ReleaseStatus.testing;
+
+        [JsonProperty("resources", NullValueHandling = NullValueHandling.Ignore)]
+        public readonly Dictionary<string, string>? Resources;
+
+        [JsonProperty("x_netkan_staging", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate,
+                                          NullValueHandling    = NullValueHandling.Ignore)]
+        [DefaultValue(false)]
+        public readonly bool Staged = false;
+
+        [JsonProperty("x_netkan_staging_reason", NullValueHandling = NullValueHandling.Ignore)]
+        public readonly string? StagingReason;
+
+        [JsonProperty("x_netkan_allow_out_of_order", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate,
+                                                     NullValueHandling    = NullValueHandling.Ignore)]
+        [DefaultValue(false)]
+        public readonly bool AllowOutOfOrder = false;
+
+        [JsonProperty("x_netkan_trust_version_file", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate,
+                                                     NullValueHandling    = NullValueHandling.Ignore)]
+        [DefaultValue(false)]
+        public readonly bool TrustVersionFile = false;
+
+        [JsonProperty("x_netkan_force_v", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate,
+                                          NullValueHandling    = NullValueHandling.Ignore)]
+        [DefaultValue(false)]
+        public readonly bool ForceV = false;
+
+        [JsonProperty("x_netkan_license_ok", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate,
+                                             NullValueHandling    = NullValueHandling.Ignore)]
+        [DefaultValue(false)]
+        public readonly bool SkipLicenseValidation = false;
+
+        [JsonProperty("x_netkan_optimus_prime", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate,
+                                                NullValueHandling    = NullValueHandling.Ignore)]
+        [DefaultValue(false)]
+        public readonly bool OptimusPrime = false;
+
+        public readonly JObject AllJson;
+
+        public JObject Json()
+            => (JObject)AllJson.DeepClone();
 
         public static Metadata Merge(Metadata[] modules)
             => //modules is [var m]
@@ -134,7 +141,7 @@ namespace CKAN.NetKAN.Model
                && modules[0] is var m
                    ? m
                    : PropertySortTransformer.SortProperties(
-                       new Metadata(MergeJson(modules.Select(m => m._json)
+                       new Metadata(MergeJson(modules.Select(m => m.AllJson)
                                                      .ToArray())));
 
         public Metadata MergeFrom(JObject[] jsons)
@@ -154,14 +161,9 @@ namespace CKAN.NetKAN.Model
 
         private static JObject MergeJson(JObject[] jsons)
         {
-            var mergeSettings = new JsonMergeSettings()
-            {
-                MergeArrayHandling     = MergeArrayHandling.Replace,
-                MergeNullValueHandling = MergeNullValueHandling.Merge,
-            };
-            var downloads = jsons.SelectMany(json => json[DownloadPropertyName] is JArray arr
+            var downloads = jsons.SelectMany(json => json["download"] is JArray arr
                                                          ? arr.Children()
-                                                         : Enumerable.Repeat(json[DownloadPropertyName], 1))
+                                                         : Enumerable.Repeat(json["download"], 1))
                                  .Distinct()
                                  .ToArray();
             var first = jsons.First();
@@ -181,45 +183,21 @@ namespace CKAN.NetKAN.Model
             // Merge game version compatibility
             GameVersion.SetJsonCompatibility(first, null, null, null);
             // Only generate array if multiple URLs
-            first[DownloadPropertyName] = downloads.Length == 1
+            first["download"] = downloads.Length == 1
                 ? downloads.First()
                 : JArray.FromObject(downloads);
             return first;
         }
 
-        public string[] Licenses
+        private static readonly JsonSerializerSettings UtcSettings = new JsonSerializerSettings()
         {
-            get
-            {
-                var lic = _json["license"];
-                switch (lic?.Type)
-                {
-                    case JTokenType.Array:
-                        return lic.Children()
-                                  .Select(t => (string?)t)
-                                  .OfType<string>()
-                                  .ToArray();
+            DateTimeZoneHandling = DateTimeZoneHandling.Utc,
+        };
 
-                    case JTokenType.String:
-                        return new string[] { (string?)lic ?? "" };
-                }
-                return Array.Empty<string>();
-            }
-        }
-
-        public bool Redistributable => Licenses.Any(lic => new License(lic).Redistributable);
-
-        public Uri? FallbackDownload
-            => Identifier != null
-            && Redistributable
-            && Version?.ToString().Replace(':', '-') is string verStr
-            && (JObject?)_json["download_hash"] is JObject hashes
-            && (string?)hashes["sha1"] is string sha1
-                ? new Uri($"https://archive.org/download/{Identifier}-{verStr}/{sha1[..8]}-{Identifier}-{verStr}.zip")
-                : null;
-
-        public bool Prerelease => _json["release_status"]?.ToString() is "testing" or "development";
-
-        public JObject Json() => (JObject)_json.DeepClone();
+        private static readonly JsonMergeSettings mergeSettings = new JsonMergeSettings()
+        {
+            MergeArrayHandling     = MergeArrayHandling.Replace,
+            MergeNullValueHandling = MergeNullValueHandling.Merge,
+        };
     }
 }